+++ administrator/components/com_crowdfunding/models/import.php
@@ -91,12 +91,12 @@ class CrowdfundingModelImport extends JModelForm
         jimport('joomla.filesystem.archive');
 
         $uploadedFile = ArrayHelper::getValue($uploadedFileData, 'tmp_name');
-        $uploadedName = ArrayHelper::getValue($uploadedFileData, 'name');
+//        $uploadedName = ArrayHelper::getValue($uploadedFileData, 'name');
         $errorCode    = ArrayHelper::getValue($uploadedFileData, 'error');
 
         // Prepare size validator.
-        $KB       = 1024 * 1024;
-        $fileSize = (int)$app->input->server->get('CONTENT_LENGTH');
+        $KB       = pow(1024, 2);
+        $fileSize = ArrayHelper::getValue($uploadedFileData, 'size', 0, 'int');
 
         $mediaParams   = JComponentHelper::getParams('com_media');
         /** @var $mediaParams Joomla\Registry\Registry */